Implement Template system
To provide a uniform style for all the pages I suggest we use a
template system. I implemented a simple template system that consist of
a header and footer. This way instead of editing ever page's style, we
only need to edit the template pages.

Easiest page to see this change is the logout.php. Also for
organization, I think we should move the css file to the template
folder, but it really doesnt matter.

If you dont think its a good idea, let me know. I can revert back.