Stylize Search Bar and Convert Search to GET
So I realized that most searches are GET methods. One big example is
google.com, when you perform a search the query is in the url. It also
gets rid of that annoying, Resend Form message. I converted the search
form to use GET instead of POST. It should work exactly the same as a
POST. POST for user login is ok.

Ill work on a homepage next.